Simplify main, process_request, run and run0
