+++ api-model/src/nitro.rs
@@ -19,7 +19,7 @@ use crate::{
 pub struct NitroEnclavesConversionRequest {
     #[cfg_attr(feature = "serde", serde(flatten))]
     pub request: ConversionRequest,
-    pub nitro_enclaves_options: NitroEnclavesConversionRequestOptions,
+    pub enclaves_options: NitroEnclavesConversionRequestOptions,
 }
 
 impl NitroEnclavesConversionRequest {
