+++ tools/container-converter/src/image_builder/snp/enclave.rs
@@ -5,15 +5,11 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use api_model::snp::SNPEnclavesMeasurements;
-use docker_image_reference::Reference as DockerReference;
-use tempfile::TempDir;
 
 use crate::Result;
 
 pub(crate) struct EnclaveImageBuilder<'a> {
-    pub(crate) client_image_reference: &'a DockerReference<'a>,
-    pub(crate) dir: &'a TempDir,
-    pub(crate) enclave_base_image: &'a DockerReference<'a>,
+    pub(crate) enclave_image_builder: crate::image_builder::enclave::EnclaveImageBuilder<'a>
 }
 impl<'a> EnclaveImageBuilder<'a> {
     // TODO: RTE-941
